refactor(helper): :truck: move common helpers to `utils` folder
