+++ src/helpers/books/audible/ApiHelper.ts
@@ -3,8 +3,8 @@ import { htmlToText } from 'html-to-text'
 import { AudibleProduct, AudibleSeries, Category } from '#config/typing/audible'
 import { ApiBook, ApiGenre, Series } from '#config/typing/books'
 import { AuthorOnBook, NarratorOnBook } from '#config/typing/people'
-import fetch from '#helpers/fetchPlus'
-import SharedHelper from '#helpers/shared'
+import fetch from '#helpers/utils/fetchPlus'
+import SharedHelper from '#helpers/utils/shared'
 import { parentCategories } from '#static/constants'
 
 class ApiHelper {
