+++ src/helpers/database/papr/audible/PaprAudibleBookHelper.ts
@@ -3,7 +3,7 @@ import { Book } from '#config/typing/books'
 import { isBook, isBookDocument } from '#config/typing/checkers'
 import { PaprBookDocumentReturn, PaprBookReturn, PaprDeleteReturn } from '#config/typing/papr'
 import { RequestGenericWithSeed } from '#config/typing/requests'
-import SharedHelper from '#helpers/shared'
+import SharedHelper from '#helpers/utils/shared'
 
 const projectionWithoutDbFields = {
 	_id: 0,
